requirements.txt
grpcio
grpcio-tools
grpcio-reflection
semantic-text-splitter==0.27.0